@@ -1,19 +1,21 @@
 {
-  "name": "Sortlog",
+  "name": "sortlog",
   "version": "1.0.0",
   "main": "dist/server.js",
   "scripts": {
     "dev": "nodemon src/server.ts --exec ts-node src/server.ts -e ts,graphql",
-    "start": "npm run build && node dist/server.js",
+    "pre-commit": "yarn prettier && yarn typecheck && yarn test",
     "build": "tsc -p . && ncp src/graphql/schemas/ dist/graphql/schemas",
-    "test": "jest --config ./jest.config.js --coverage --watchAll --testPathPattern=test --detectOpenHandles",
-    "doc": "node src/utils/swagger.ts"
+    "start": "node dist/server.js",
+    "prettier": "prettier -c .prettierrc --write",
+    "test": "jest --colors --coverage --testPathPattern=test --detectOpenHandles",
+    "test:watch": "jest --colors --coverage  --watchAll --testPathPattern=test --detectOpenHandles",
+    "typecheck": "tsc --project tsconfig.json --pretty --noEmit"
   },
   "dependencies": {
     "apollo-server": "^3.6.3",
     "apollo-server-express": "^2.26.0",
     "apollo-server-testing": "^2.25.3",
-    "chai": "^4.3.6",
     "compression": "^1.7.4",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
@@ -24,38 +26,31 @@
     "graphql-import-node": "^0.0.1",
     "http": "^0.0.0",
     "http-status-codes": "^2.2.0",
-    "jest": "^27.5.1",
-    "jest-transform-graphql": "^2.1.0",
     "jsonwebtoken": "^8.5.1",
-    "mocha": "^9.2.1",
     "mongoose": "^5.10.6",
-    "ncp": "^2.0.0",
     "swagger-jsdoc": "^6.2.5",
     "swagger-ui-express": "^4.5.0",
-    "ts-jest": "^27.1.3",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.7",
-    "xss-clean": "^0.1.1"
+    "xss-clean": "^0.1.1",
+    "typescript": "^4.8.4"
   },
   "devDependencies": {
     "@types/compression": "^0.0.36",
     "@types/express": "^4.17.8",
     "@types/graphql": "^14.5.0",
     "@types/graphql-depth-limit": "^1.1.2",
+    "@types/jest": "^29.1.1",
     "@types/jsonwebtoken": "^8.5.0",
-    "@types/mocha": "^9.1.0",
     "@types/mongoose": "^5.7.36",
-    "@types/node": "^11.15.27",
+    "@types/node": "^18.7.23",
     "@types/swagger-jsdoc": "^6.0.1",
     "@types/swagger-ui-express": "^4.1.3",
-    "nodemon": "^1.19.4"
-  },
-  "jest": {
-    "preset": "ts-jest",
-    "globals": {
-      "ts-jest": {
-        "tsConfig": "tsconfig.json"
-      }
-    }
+    "chai": "^4.3.6",
+    "jest": "^29.1.1",
+    "jest-transform-graphql": "^2.1.0",
+    "ncp": "^2.0.0",
+    "nodemon": "^1.19.4",
+    "prettier": "^2.6.2",
+    "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1"
   }
 }
